+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Get user profile when retrieve just the user's e-mail.
+- Redirect to the return URL when go to login page and the it's already logged.
 
 ## [1.5.5] - 2018-09-26
 ### Fixed