+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - Get user profile when retrieve just the user's e-mail.
-- Redirect to the return URL when go to login page and the it's already logged.
+- Redirect to the return URL when go to the Login page, but it's already logged.
 
 ## [1.5.5] - 2018-09-26
 ### Fixed